# docs/openapi.yml
openapi: 3.0.0
servers:
  - description: Production
    url: https://localhost:9090
info:
  description: Codewind API
  version: 0.1.0
  title: Codewind API
  contact:
    name: API support
    email: example@ibm.com
  license:
    name: Microclimate Developer License
    url: 'https://github.ibm.com/dev-ex/codewind/blob/master/LICENSE/LICENSE.md'

# tags:
  # - name: users
  #   description: Operations available to regular users
  # - name: admin
  #   description: Operations available only to authenticated admins
paths:
...
  /api/v1/projects/bind:
    post:
      summary: Bind a project to Codewind
      description: >
        The bind api will be passed the location on disk of an existing project so that Codewind
        adds it to the list of projects it monitors.
      requestBody:
        description: project bind options
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - path
                - language
                - projectType
              properties:
                name:
                  $ref: '#/components/schemas/ProjectName'
                path:
                  description: the absolute path on the file system to the project
                  type: string
                  example: /Documents/projectDir/
                language:
                  $ref: '#/components/schemas/ProjectLanguage'
                projectType:
                  $ref: '#/components/schemas/ProjectType'
                autoBuild:
                  description: whether the project should be built automatically (default true)
                  type: boolean
      responses:
        202:
          description: >
            Bind request accepted. Will attempt to bind the project directory as a Codewind project.
            When binding is complete the projectBind socket event will be sent. If autoBuild was true
            then this will be sent after the build is complete, otherwise it will be sent straight away.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        400:
          $ref: '#/components/responses/400'
        409:
          description: The project path or name are already in use
        500:
          description: Any other error
...
  /api/v1/projects/{id}/unbind:
      post:
        summary: Unbind a given project from codewind
        description: >
          This requests Codewind unbinds the project location for the project with the specified id.
          Codewind will emit a `projectDeletion` event on completion.
        parameters:
          - in: path
            name: id
            schema:
              type: string
            required: true
            description: id of project
        responses:
          202:
            description: if project deletion was accepted
          404:
            description: if the project with id was not found
          409:
            description: if unbind was already in progress
...
  /api/v1/projects/{id}/file-changes/{projectWatchStateId}/status:
    put:
      summary: Report the project watch state for a praticular projectWatchStateId
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
          description: id of project
        - name: projectWatchStateId
          in: path
          schema:
            type: string
          required: true
          description: projectWatchStateId represents a file watching client's current state. projectWatchStateId is generated and returned when first time calling the endpoint `GET /api/v1/projects/watchlist`. Additionally, if the ignored paths list for a project is changed, a new projectWatchStateId is generated and sent to file watchers via WebSocket event from `/websockets/file-changes/v1`
        - name: clientUUID
          in: query
          schema:
            type: string
            format: uuid
          required: true
          description: a unique id of the client
      
      responses:
        200:
          description: if successfully received the request
        400:
          $ref: '#/components/responses/400'
        404:
          description: The project was not found
        500:
          description: if an error occured
  /api/v1/projects/:id/file-changes:
    post:
      summary: Notify the event occurrance with a list of changed files of the project
      parameters:
        - name: id
          in: path
          schema:
            type: string
          required: true
          description: id of project
        - name: timestamp
          in: query
          schema:
            type: number
          required: true
          description: timestamp of this file change event
        - name: chunk
          in: query
          schema:
            type: number
          required: true
          description: chunk number of this current chunk
        - name: chunk_total
          in: query
          schema:
            type: number
          required: true
          description: total chunks expected for this timestamp
      requestBody:
        description: A JSON array of which file paths have changed, which is then compressed with zlib, then converted to base64. The example shows JSON content before converted to base64
        content:
          application/json:
            schema:
              type: object
              required:
                - msg
              properties:
                msg:
                  type: array
                  items:
                    $ref: '#/components/schemas/FileChangeList'
      responses:
        200:
          description: if successfully received the request
        400:
          $ref: '#/components/responses/400'
        404:
          description: The project was not found
        500:
          description: if an error occured
...
components:
  schemas:
    Project:
      type: object
      required:
        - projectID
        - name
        - language
        - projectType
        - host
        - ports
      properties:
        projectID:
          $ref: '#/components/schemas/ProjectID'
        name:
          $ref: '#/components/schemas/ProjectName'
        language:
          $ref: '#/components/schemas/ProjectLanguage'
        projectType:
          $ref: '#/components/schemas/ProjectType'
        host:
          type: integer
          example: localhost
        ports:
          type: integer
          example: 31234
    ProjectID:
      type: string
      format: uuid
      example: d290f1ee-6c54-4b01-90e6-d701748f0851
      pattern: '[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}'
    ProjectName:
      description: Name of project, cannot contain any of ", &, /, \\, ?, %, *, :, |, <, > or spaces
      type: string
      example: tobesproject
    ProjectLanguage:
      type: string
    ProjectType:
      description: the build type for the project
      type: string
      enum: [ 'liberty', 'spring', 'swift', 'nodejs', 'docker' ]
...
    FileChangeList: 
        type: object
        required: 
          - path
          - timestamp
          - type
          - directory
        properties:
          path:
            description: the absolute path on the file system to the file
            type: string
            example: Documents/projectDir/File1
          timestamp:
            description: the timestamp of this change
            type: number
            example: 2117514496
          type:
            description: the change type, one of the following CREATE / MODIFY / DELETE
            type: string
            example: CREATE
          directory:
            description: true or false, on whether the path refers to a directory
            type: boolean
            example: false
...
  responses:
    400:
      description: Bad request
      content:
        application/json:
          schema:
            type: string
...